move app into page
